+++ sdk/rdbms/azure-mgmt-rdbms/generated_samples/mysql_flexibleservers/operation_results_get.py
@@ -14,7 +14,7 @@ from azure.mgmt.rdbms.mysql_flexibleservers import MySQLManagementClient
     pip install azure-identity
     pip install azure-mgmt-rdbms
 # USAGE
-    python server_reset_gtid.py
+    python operation_results_get.py
 
     Before run the sample, please set the values of the client ID, tenant ID and client secret
     of the AAD application as environment variables: AZURE_CLIENT_ID, AZURE_TENANT_ID,
@@ -26,16 +26,16 @@ from azure.mgmt.rdbms.mysql_flexibleservers import MySQLManagementClient
 def main():
     client = MySQLManagementClient(
         credential=DefaultAzureCredential(),
-        subscription_id="ffffffff-ffff-ffff-ffff-ffffffffffff",
+        subscription_id="11111111-2222-3333-4444-555555555555",
     )
 
-    client.servers.begin_reset_gtid(
-        resource_group_name="TestGroup",
-        server_name="testserver",
-        parameters={"gtidSet": "4aff5b51-97ba-11ed-a955-002248036acc:1-16"},
-    ).result()
+    response = client.operation_results.get(
+        location_name="westus",
+        operation_id="resource-provisioning-id-farmBeatsResourceName",
+    )
+    print(response)
 
 
-# x-ms-original-file: specification/mysql/resource-manager/Microsoft.DBforMySQL/FlexibleServers/preview/2023-10-01-preview/examples/ServerResetGtid.json
+# x-ms-original-file: specification/mysql/resource-manager/Microsoft.DBforMySQL/ServiceOperations/preview/2023-06-01-preview/examples/OperationResults_Get.json
 if __name__ == "__main__":
     main()